requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==5.2.3
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-crispy-forms==2.4
django-js-asset==3.1.2
django-taggit==6.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
